+++ packages/browser/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @glissade/browser
 
+## 0.60.0-pre.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @glissade/scene@0.60.0-pre.1
+  - @glissade/backend-canvas2d@0.60.0-pre.1
+  - @glissade/backend-dom@0.60.0-pre.1
+  - @glissade/element@0.60.0-pre.1
+  - @glissade/player@0.60.0-pre.1
+  - @glissade/core@0.60.0-pre.1
+
 ## 0.60.0-pre.0
 
 ### Patch Changes